+++ src/main/java/kitchenpos/application/MenuService.java
@@ -6,11 +6,11 @@ import kitchenpos.application.exception.MenuGroupNotFoundException;
 import kitchenpos.application.exception.ProductNotFoundException;
 import kitchenpos.domain.menu.repository.MenuRepository;
 import kitchenpos.repository.MenuGroupRepository;
-import kitchenpos.repository.ProductRepository;
+import kitchenpos.domain.product.repository.ProductRepository;
 import kitchenpos.domain.menu.Menu;
 import kitchenpos.domain.MenuGroup;
 import kitchenpos.domain.menu.MenuProduct;
-import kitchenpos.domain.Product;
+import kitchenpos.domain.product.Product;
 import kitchenpos.ui.dto.request.CreateMenuProductRequest;
 import kitchenpos.ui.dto.request.CreateMenuRequest;
 import org.springframework.stereotype.Service;
