+++ src/main/java/kitchenpos/ui/dto/response/ReadProductResponse.java
@@ -1,7 +1,7 @@
 package kitchenpos.ui.dto.response;
 
 import java.math.BigDecimal;
-import kitchenpos.domain.Product;
+import kitchenpos.domain.product.Product;
 
 public class ReadProductResponse {
