+++ src/test/java/kitchenpos/application/MenuServiceTest.java
@@ -11,11 +11,11 @@ import kitchenpos.config.IntegrationTest;
 import kitchenpos.domain.menu.Menu;
 import kitchenpos.domain.MenuGroup;
 import kitchenpos.domain.menu.MenuProduct;
-import kitchenpos.domain.Product;
+import kitchenpos.domain.product.Product;
 import kitchenpos.domain.exception.InvalidPriceException;
 import kitchenpos.repository.MenuGroupRepository;
 import kitchenpos.domain.menu.repository.MenuRepository;
-import kitchenpos.repository.ProductRepository;
+import kitchenpos.domain.product.repository.ProductRepository;
 import kitchenpos.ui.dto.request.CreateMenuProductRequest;
 import kitchenpos.ui.dto.request.CreateMenuRequest;
 import org.junit.jupiter.api.Test;
