+++ src/test/java/kitchenpos/domain/OrderLineItemsTest.java
@@ -11,6 +11,7 @@ import java.util.List;
 import kitchenpos.domain.exception.InvalidOrderLineItemException;
 import kitchenpos.domain.menu.Menu;
 import kitchenpos.domain.menu.MenuProduct;
+import kitchenpos.domain.product.Product;
 import org.junit.jupiter.api.DisplayNameGeneration;
 import org.junit.jupiter.api.DisplayNameGenerator;
 import org.junit.jupiter.api.Test;
